Initial commit of application code
First commit of code for the FastAPI python webapp.

Includes a "method" module that contains many of the functions used in the main.py file.

styles.css is a generic style sheet for the output table.

table.html is a Jinja2 template that displays the card data as a dynamic HTML table